Modify aggregates so they are useable for true cross publishing
There is a requirement to be able to copy packages across repositories
of different architecture. This is useful for packages that need to be
in sync across those repositories or to inject cross compilers and
sb2 acceleration tools.

Currently the scheduler refuses to operate on any package unless it is
built for own arch. This patch makes this configurable by allowing an arch
element in the aggregate's xml schema. The values of the arch elements
are used to look at other arch builds of the aggregate source.

If the target of the aggregate produces noarch binaries then the rpms are
usable immediately once the aggregation is done.
Otherwise using baselibs.conf is required to convert the arch dependant binaries.

Possibly solved use cases:
* synchronized versioning packages across different repositories
* easier "injection" of sb2 tools